requirements.txt
pytz
mysql-connector-python
